pyproject.toml
[project]
name = "datadex"
version = "1.0.0"
authors = [{ name = "David Gasquez" }]
dependencies = [
    "dagster-dbt>=0.24.13",
    "dagster-duckdb>=0.24.13",
    "dagster-duckdb-polars>=0.24.13",
    "dagster>=1.8.13",
    "datasets>=3.0.2",
    "dbt-core>=1.8.8",
    "dbt-duckdb>=1.9.0",
    "duckdb>=1.1.2",
    "httpx[http2]>=0.27.2",
    "huggingface-hub>=0.26.1",
    "ipykernel>=6.29.5",
    "pyarrow>=17.0.0",
    "python-slugify>=8.0.4",
    "tenacity>=9.0.0",
]

requires-python = ">=3.11, <=3.13"
readme = "README.md"
license = { text = "MIT" }

[project.urls]
Homepage = "https://datadex.datonic.io/"
Changelog = "https://github.com/datonic/datadex/commits/main/"
Issues = "https://github.com/datonic/datadex/issues"
CI = "https://github.com/datonic/datadex/actions"

[tool.setuptools]
packages = ["datadex"]

[tool.dagster]
module_name = "datadex.definitions"
code_location_name = "datadex"

[tool.uv]
dev-dependencies = ["dagster-webserver>=1.8.13"]
